# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DTuin Faculty Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  Normal: "Normal",
  Alert: "Alert",
  Warning: "Warning",
} as const;

export interface Room {
  id: number;
  name: string;
  floor: number;
  area: number;
  capacity: number;
  students: number;
  temperature: number;
  energyUsage: number;
  status: RoomStatus;
  occupancyRate: number;
}

export interface AnalyticsSummary {
  totalRooms: number;
  totalStudents: number;
  energyUsage: number;
  occupancyRate: number;
  avgTemperature: number;
  activeAlerts: number;
}

export interface OccupancyData {
  room: string;
  occupancy: number;
}

export interface EnergyTrend {
  month: string;
  energy: number;
}

export interface TemperatureData {
  room: number;
  floor: number;
  temperature: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  red: "red",
  yellow: "yellow",
  green: "green",
} as const;

export interface Alert {
  id: number;
  roomName: string;
  type: string;
  message: string;
  severity: AlertSeverity;
  timestamp: string;
}