# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DTuin Faculty Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{id}:
    get:
      operationId: getRoomById
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        floor:
          type: integer
        area:
          type: number
        capacity:
          type: integer
        students:
          type: integer
        temperature:
          type: number
        energyUsage:
          type: number
        status:
          type: string
          enum: [Normal, Alert, Warning]
        occupancyRate:
          type: number
      required:
        - id
        - name
        - floor
        - area
        - capacity
        - students
        - temperature
        - energyUsage
        - status
        - occupancyRate
...