# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DTuin Faculty Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from "./roomStatus";

export interface Room {
  id: number;
  name: string;
  floor: number;
  area: number;
  capacity: number;
  students: number;
  temperature: number;
  energyUsage: number;
  status: RoomStatus;
  occupancyRate: number;
}